# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ремонт часов на Алексеевской",
        "contact": {
            "name": "Анастасия Паршенкова",
            "email": "parshenkova.anastasya@yandex.ru"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/orders/": {
            "get": {
                "tags": [
                    "Order"
                ],
                "description": "Возвращает перечень заказов",
                "operationId": "acf76e3d313115c8d911054877dec32e",
                "responses": {
                    "200": {
                        "description": "Перечень заказов",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "description": "Создаёт заказ",
                "operationId": "69c0ceaab14d4413bc26b6616d0e9110",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заказ создан",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{order}/": {
            "get": {
                "tags": [
                    "Order"
                ],
                "description": "Возвращает заказ по идентификатору",
                "operationId": "0ced17f2d4e76d75e39b6e4e5227ad56",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/order"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заказ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order"
                ],
                "description": "Удаляет запись о заказе",
                "operationId": "452181cb5ee4d5740a6c437078b71437",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/order"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Запись о заказе удалена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Order"
                ],
                "description": "Корректирует запись о заказе",
                "operationId": "2efc465e92b1f73d8f899f8c7d6c0a25",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/order"
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Запись о заказе откорректирована",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/order_product/": {
            "get": {
                "tags": [
                    "OrderProduct"
                ],
                "description": "Возвращает перечень закзанных продуктов",
                "operationId": "3742f71e69b523673850b4dc365b9647",
                "responses": {
                    "200": {
                        "description": "Перечень закзанных продуктов",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Закзанный продукт не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "OrderProduct"
                ],
                "description": "Создаёт закзанный продукт",
                "operationId": "7d0efcf81b253d13aea707f037b24d9a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderProduct"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Закзанный продукт создан",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/order_product/{order_product}/": {
            "get": {
                "tags": [
                    "OrderProduct"
                ],
                "description": "Возвращает закзанный продукт по идентификатору",
                "operationId": "69b555cf527dd189e131df2e73021340",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/order_product"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Закзанный продукт",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Закзанный продукт не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "OrderProduct"
                ],
                "description": "Удаляет запись о заказанном продукте",
                "operationId": "cfa459d168784d2f5ee64c182d2f1199",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/order_product"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Запись о заказанном продукте удалена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "OrderProduct"
                ],
                "description": "Корректирует запись о заказанном продукте",
                "operationId": "47cd4678c8973797fba0999517d8e370",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/order_product"
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderProduct"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Запись о заказанном продукте откорректирована",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Order": {
                "title": "Order",
                "required": [
                    "total_price",
                    "user_id"
                ],
                "properties": {
                    "total_price": {
                        "description": "Общая цена",
                        "type": "integer",
                        "format": "float"
                    },
                    "user_id": {
                        "description": "ID пользователя",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "OrderProduct": {
                "title": "OrderProduct",
                "required": [
                    "quantity",
                    "order_id",
                    "product_id"
                ],
                "properties": {
                    "quantity": {
                        "description": "Количество",
                        "type": "integer"
                    },
                    "order_id": {
                        "description": "ID заказа",
                        "type": "integer"
                    },
                    "product_id": {
                        "description": "ID продукта",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "order": {
                "name": "order",
                "in": "path",
                "description": "Первичный ключ"
            },
            "order_product": {
                "name": "order_product",
                "in": "path",
                "description": "Первичный ключ"
            },
...
        }
    },
...
}